+++ CHANGELOG.md
@@ -18,6 +18,24 @@ Date format: `YYYY-MM-DD`
 ### Security
 
 ---
+## [1.1.0] - 2024-10-25
+
+### Added
+- **FEATURE:** Added support for a `Parsing` interface to allow for custom version parsing.
+- **FEATURE:** Added documentation for the configuration options: `WithStrictAdherence`.
+
+### Changed
+- **DEBT:** Added additional functional tests to improve code coverage.
+- **DEBT:** Refactored the benchmark tests to correctly measure performance.
+
+### Deprecated
+### Removed
+### Fixed
+- **DEFECT:** Corrected various documentation issues.
+
+### Security
+
+---
 ## [1.0.0] - 2024-10-24
 
 ### Added
@@ -28,7 +46,8 @@ Date format: `YYYY-MM-DD`
 ### Fixed
 ### Security
 
-[Unreleased]: https://github.com/sixafter/semver/compare/v1.0.0...HEAD
+[Unreleased]: https://github.com/sixafter/semver/compare/v1.1.0...HEAD
+[1.0.0]: https://github.com/sixafter/semver/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/sixafter/semver/compare/d63ed577e7e841fb2209cfdcf4269fac6a57c85e...v1.0.0
 
 [MUST]: https://datatracker.ietf.org/doc/html/rfc2119